requirements.txt
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.1.1
cycler==0.12.1
fonttools==4.43.1
idna==3.6
kiwisolver==1.4.5
matplotlib==3.8.0
MouseInfo==0.1.3
networkx==3.2.1
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
PyAudio==0.2.14
PyAutoGUI==0.9.54
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.8.2
python3-xlib==0.15
pytweening==1.0.7
requests==2.31.0
six==1.16.0
SpeechRecognition==3.10.0
urllib3==2.1.0
